Description of files
AbstractCompetitor Class:

It is an abstract class serving as a base class for competitors.
Annotated with @Entity indicates that instances of this class can be persisted to a database.
Contains an @Id annotated field id with a @SequenceGenerator and @GeneratedValue annotations. This means that the id field will be automatically generated using a sequence named "competitor_sequence" when a new instance is persisted.
Contains attributes name (String) and rank (double) with their respective getter and setter methods.
Has a default constructor that generates a random UUID for the id.
Declares an abstract method performAction(), which must be implemented by subclasses.
Player Class:

Extends the AbstractCompetitor class, inheriting its attributes and methods.
Annotated with @Table(name = "player") specifies the table name as "player" in the database.
Overrides the performAction() method with a specific implementation.
May have additional fields and methods specific to the Player class.
JPA Annotations:

@Entity: Marks the class as a JPA entity, allowing instances to be persisted to a database.
@Table(name = "player"): Specifies the table name in the database. In this case, it is set to "player."
@Id: Marks the field as the primary key.
@SequenceGenerator: Specifies a database sequence generator named "competitor_sequence" with an allocation size of 1.
@GeneratedValue: Specifies that the id field should be generated using the specified sequence generator.
UUID Generation:

The AbstractCompetitor class uses UUID.randomUUID() in its default constructor to generate a random UUID for the id field.
Abstract Methods:

The AbstractCompetitor class declares an abstract method performAction(), which must be implemented by subclasses. This allows for specific behavior to be defined in each subclass.
Overall, this code provides a flexible and extensible framework for modeling competitors in a system, with the ability to persist instances to a database using JPA. Subclasses, such as the Player class, can extend the base functionality and define specific behaviors.